1071.greatest_common_divisor_of_strings/main.go
package main

// use recursion
// time complexity: O(m+n)
// space complexity: O(m+n)
func gcdOfStrings(str1, str2 string) string {
	if str1 == str2 {
		return str1
	}

	if len(str2) > len(str1) {
		str1, str2 = str2, str1
	}

	if str1[:len(str2)] != str2 {
		return ""
	}

	return gcdOfStrings(str1[len(str2):], str2)
}

// time complexity: O(n)
// space complexity: O(1)
func gcdOfStrings2(str1, str2 string) string {
	if str1+str2 != str2+str1 {
		return ""
	}
	return str1[:gcd(len(str1), len(str2))]
}

func gcd(a, b int) int {
	for b != 0 {
		a, b = b, a%b
	}
	return a
}
